RadioGroup props interface localized
