+++ src/__stories/inputs/radio.stories.tsx
@@ -1,31 +1,19 @@
 import React, { ReactElement } from "react";
-import RadioGroup from "../../__ui/formElements/radiogroup/radio-group";
-import Radio from "../../__ui/formElements/radiogroup/radio";
+import { RadioGroup, RadioItem } from "../../__ui/formElements/radiogroup";
+
 import StoryPage from "../story-page";
-import { lorem, random } from "faker";
+import { lorem } from "faker";
 
 export default {
   title: "Form Elements/Radio Group",
   component: RadioGroup,
+  subcomponents: { RadioItem },
   parameters: {
     componentSubtitle: "Custom Radio group component",
   },
 };
 
-const radioGroup = (({ len }): ReactElement[] => {
-  const seletcedIndex = random.number({ min: 0, max: len });
-  return new Array(len).fill("").map(
-    (chk: string, i: number): ReactElement => {
-      return (
-        <Radio key={i} checked={i === seletcedIndex} label={lorem.words(4)} value={`ABC-${i}`}>
-          <span>{lorem.paragraph()}</span>
-        </Radio>
-      );
-    }
-  );
-})({ len: 3 });
-
-export const Primary = function(): ReactElement {
+export const Primary = function (): ReactElement {
   return (
     <StoryPage viewType="PrimaryView">
       <RadioGroup
@@ -33,21 +21,22 @@ export const Primary = function(): ReactElement {
           console.log("PrimaryView RadioGroup[1] value", val);
         }}
       >
-        {radioGroup}
+        <RadioItem checked={true} label={lorem.words(4)} value={`ABC-1`}>
+          <span>{lorem.paragraph()}</span>
+        </RadioItem>
+        <RadioItem label={lorem.words(4)} value={`ABC-2`}>
+          <span>{lorem.paragraph()}</span>
+        </RadioItem>
+        <RadioItem label={lorem.words(4)} value={`ABC-3`}>
+          <span>{lorem.paragraph()}</span>
+        </RadioItem>
       </RadioGroup>
       <hr />
-      <RadioGroup
-        onChange={(val: string | number): void => {
-          console.log("PrimaryView RadioGroup[2] value", val);
-        }}
-      >
-        {radioGroup}
-      </RadioGroup>
     </StoryPage>
   );
 };
 
-export const Secondary = function(): ReactElement {
+export const Secondary = function (): ReactElement {
   return (
     <StoryPage viewType="SecondaryView">
       <RadioGroup
@@ -55,7 +44,15 @@ export const Secondary = function(): ReactElement {
           console.log("SecondaryView RadioGroup[1] value", val);
         }}
       >
-        {radioGroup}
+        <RadioItem checked={true} label={lorem.words(4)} value={`ABC-1`}>
+          <span>{lorem.paragraph()}</span>
+        </RadioItem>
+        <RadioItem label={lorem.words(4)} value={`ABC-2`}>
+          <span>{lorem.paragraph()}</span>
+        </RadioItem>
+        <RadioItem label={lorem.words(4)} value={`ABC-3`}>
+          <span>{lorem.paragraph()}</span>
+        </RadioItem>
       </RadioGroup>
     </StoryPage>
   );
