+++ src/__ui/formElements/index.tsx
@@ -2,8 +2,8 @@ import Input from "./input";
 import MultipleInput from "./multipleInput";
 import Checkbox from "./checkbox/checkbox";
 import RadioGroup from "./radiogroup/radio-group";
-import Radio from "./radiogroup/radio";
+import RadioItem from "./radiogroup/radio-item";
 import DateTime from "./dateTime";
 import Suggestion from "./suggestion";
 
-export { Input as default, MultipleInput, DateTime, Checkbox, RadioGroup, Radio, Suggestion };
+export { Input as default, MultipleInput, DateTime, Checkbox, RadioGroup, RadioItem as Radio, Suggestion };
