+++ src/__ui/formElements/radiogroup/radio-item.tsx
@@ -1,10 +1,22 @@
-import React, { FunctionComponent } from "react";
-import { IRadioItemProps, IconSizeEnums, IconNameEnums, ThemeEnums } from "../../../__typings/interfaces.d";
+import React, { FunctionComponent, ReactNode } from "react";
+import { IconSizeEnums, IconNameEnums, ThemeEnums } from "../../../__typings/interfaces.d";
 import themeDefault from "./theme-default.module.scss";
 import themeOcean from "./theme-ocean.module.scss";
 import { useTheme, useThemeStyle } from "../../typography";
 
-import Icon from "../../../__ui/icon";
+import Icon from "../../icon";
+
+export interface IRadioItemProps {
+  children?: ReactNode;
+  /** *initial state to be set* */
+  checked?: boolean;
+  /** label text identifies the checkbox or radio element */
+  label: string;
+  /** value of the radio element, whihc will be assigend to radiogroup when it's selected */
+  value: string;
+  /** callback function to be don if checkbox or radio  changes it state */
+  onChange?: (val?: string) => void;
+}
 
 const stylesMap = new Map();
 stylesMap.set(ThemeEnums.OCEAN_THEME, themeOcean);
@@ -15,13 +27,13 @@ stylesMap.set(ThemeEnums.DEFAULT_THEME, themeDefault);
  *
  */
 
-const Radio: FunctionComponent<IRadioItemProps> = ({
+export const RadioItem: FunctionComponent<IRadioItemProps> = ({
   children,
   label = "",
   onChange = (): boolean => false,
   checked = false,
   value,
-}) => {
+}: IRadioItemProps) => {
   const theme = useTheme();
   const styles = useThemeStyle(stylesMap);
 
@@ -36,5 +48,3 @@ const Radio: FunctionComponent<IRadioItemProps> = ({
     </>
   );
 };
-
-export default Radio;
